Stage 5-D: the 1.20.1 rig, and the five things only a running server could say
The node built and mixin-checked clean at the end of Stage 5-C. It did not boot.
This is the rig that asked, and the five answers.

The rig: scratchpad/smoke-1.20.1 (fabric server 1.20.1 / loader 0.16.10,
fabric-api 0.92.11, Skill Proficiencies 1.6.0+1.20.1, no PAL per Option B),
driver arch-gate/smoke1201.sh on Java 17. eula.txt copied from the 1.21.1 rig.
Three probe families, each with a bogus control, because all three failure modes
are silent on a dedicated server.

Four bugs it found in the first twenty minutes, none of which a build can see:

* PlayerMixin's legacy knockback stash had no `DamageSource` local to bind.
  1.20.1 builds the source inline at both sites instead of hoisting it. Stage 4
  predicted a failure here and predicted the wrong cause: the sweep is spelled
  identically (two knockback(DDD) at 479/714, both inside the hurt branch), the
  SOURCE is what differs. Narrowed to >=1.21; new <1.21 arm rebuilds the source
  with vanilla's own expression.
* AbstractArrowAccessor died applying accessors: `pickupItemStack` is a field
  from 1.21 and an abstract getter below it. remapJar had said so as a WARNING
  and shipped the jar anyway. Now an @Invoker below 1.21.
* archetypes$spellbowGravity did not resolve: the constant is not 0.05, it is
  0.05000000074505806 — a widened float. Spelled `(double) 0.05F`.
* R-B9, wider than filed: the datapack registry directories are plural below
  1.21, and relocating them exposed that the recipe SCHEMA breaks on 1.21.1 too.
  That node's Stage-4 smoke had been logging twenty parse errors the gate never
  grepped for. Per-node overrides for both legacy nodes, README'd beside the
  files; breeze_wand excised on 1.20.1 (no breeze_rod there, and substituting a
  different ingredient is a balance divergence).

And the fifth, which the gate found rather than the boot: R-B1's client half had
never been written. LegacyStateSync encodes, routes, replays on start-tracking
and replays on join, and nothing anywhere registered to receive the channel.
No error exists for this — server state stays perfect and the entire visible half
is blank on one node. client/LegacyStateSyncClient closes it.

Gates 2/3/5/6 green on 1.20.1: boots and stops clean, exit 0, 17/19 common mixins
(the two absent have no player-free target class), zero injection failures, zero
tag failures, zero parse errors, and the funnel is the 1.21.1 sequence plus both
mods' AFTER_DAMAGE substitutes, landing at 500 < 900 < 1000 on every RETURN.
Flense still last after Skill Proficiencies' three. mixincheck ALL RESOLVE on all
five nodes.

Also here: Swift Shadow's client arm (@ModifyArg on the Mth.clamp aiStep feeds to
Input.tick, so vanilla's own clamp is what caps it at 1.0 — the same [0,1] the
SNEAKING_SPEED attribute's range gives on every other node), and the meat tag
written out on 1.20.1 where #minecraft:meat does not exist.

R-C2 closed with no change: all eight damage types exist on both legacy nodes.

Prior-node identity: 26.2 244/244, 26.1 244/244, 1.21.11 250/250 classes and
385/385 resources unmoved. 1.21.1 250/250 classes, +0/-0 shape, exactly the 20
recipe resources changed.

+++ src/main/java/com/archetypes/mixin/AbstractArrowMixin.java
@@ -66,11 +66,27 @@ public abstract class AbstractArrowMixin {
 	// `if (!this.isNoGravity())` — the same short-circuit the modern hook composes with, so a
 	// True Shot's no-gravity arrow still ignores both marks in the same order.
 	//
-	// MEASURED, not assumed (`javap -c` on the 1.20.1 mojmap jar): `tick()` contains exactly
-	// ONE `double 0.05` (offset 859, the `Vec3.y` subtraction) and one unrelated `float
-	// 0.05f` (offset 739, the water/air inertia pair) which a doubleValue constant cannot
-	// match. So the injection point is unambiguous, and the product it produces —
-	// `0.05 * SPELLBOW_ARROW_GRAVITY_FACTOR` — is the same number the modern arm returns.
+	// MEASURED (`javap -c -constants` on the 1.20.1 mojmap jar): `tick()` guards the
+	// subtraction with `832: isNoGravity()` / `835: ifne`, then does
+	// `setDeltaMovement(v.x, v.y - <const>, v.z)` across offsets 848-868 — vanilla's own
+	// short-circuit, exactly as the modern arm composes with.
+	//
+	// AND THE CONSTANT IS NOT `0.05`. Stage 5 wrote `doubleValue = 0.05` from a reading of
+	// that javap dump that stopped at the third character, and `injectors.defaultRequire: 1`
+	// duly failed the node's boot with `expected 1 invocation(s) but 0 succeeded`. The
+	// literal at offset 859 is `ldc2_w double 0.05000000074505806d` — vanilla writes the
+	// gravity as a FLOAT and javac widens it, so the double in the constant pool is
+	// `(double) 0.05F` and nothing else. Spelled that way here rather than as the 17 digits,
+	// because the digits are what invites the same mistake back: this is a widened float, and
+	// the annotation's own widening produces the identical bits.
+	//
+	// Nothing else in `tick()` can be caught by it: the only other 0.05 in the method is the
+	// unrelated `float 0.05f` at 739 (the water/air inertia pair), which a `doubleValue`
+	// constant cannot match, and no other `ldc2_w` in the method carries this value.
+	//
+	// The product it produces — `(double) 0.05F * SPELLBOW_ARROW_GRAVITY_FACTOR` — is the
+	// same number the modern arm returns, because `getDefaultGravity()` returns that same
+	// widened float from 1.20.5 up.
 	//? if >=1.20.5 {
 	@ModifyReturnValue(method = "getDefaultGravity()D", at = @At("RETURN"))
 	private double archetypes$spellbowGravity(final double original) {
@@ -78,7 +94,7 @@ public abstract class AbstractArrowMixin {
 	}
 	//?} else {
 	/*@org.spongepowered.asm.mixin.injection.ModifyConstant(method = "tick()V",
-			constant = @org.spongepowered.asm.mixin.injection.Constant(doubleValue = 0.05))
+			constant = @org.spongepowered.asm.mixin.injection.Constant(doubleValue = (double) 0.05F))
 	private double archetypes$spellbowGravity(final double original) {
 		return archetypes$spellbowGravityImpl(original);
 	}
